Move enocean base entity to separate module
