+++ homeassistant/components/enocean/entity.py
